requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
pillow==10.1.0
openai-whisper